gosps: move GosPackageStatePm to com.android.server package
It's supposed to be used only within system_server.
